Module Attr_helper

Helpers for attributes

Warning: this module is unstable and part of compiler-libs.

type error =

| Multiple_attributes of string

| No_payload_expected of string

val get_no_payload_attribute : string list -> Parsetree.attributes -> string Asttypes.loc option

The string list argument of the following functions is a list of alternative names for the attribute we are looking for. For instance:

["foo"; "ocaml.foo"]

val has_no_payload_attribute : string list -> Parsetree.attributes -> bool

exception Error of Location.t * error

val report_error : Stdlib.Format.formatter -> error -> unit