Module Bitwuzla_z

val value_get_bv_bits : Bitwuzla_c.term -> Z.t

value_get_bv_bits term get the bits of a given term.

parameter term

The value term.

returns

The unsigned value of term as a Zarith integer.

raises Invalid_argument

if the term is not a value.