Module Closure_conversion_aux.Function_decls

Used to represent information about a set of function declarations during closure conversion. (The only case in which such a set may contain more than one declaration is when processing "let rec".)

module Function_decl : sig ... end

type t

val create : Function_decl.t list -> t

val to_list : t -> Function_decl.t list

val all_free_idents : t -> Ident.Set.t

val closure_env_without_parameters : Env.t -> t -> Env.t