Module Cmm_invariants

Check a number of continuation-related invariants

val run : Stdlib.Format.formatter -> Cmm.fundecl -> bool

run ppf fundecl analyses the given function, and returns whether any errors were encountered (with corresponding error messages printed on the given formatter).