Module 1-S.Subrange_state

The module Subrange_state describes information that needs to be propagated and passed to Subrange_info.create. The state that will be used for subrange creation is the state at the end of the subrange, not at the beginning.

type t

val create : unit -> t

val advance_over_instruction : t -> Compute_ranges_intf.L.instruction -> t