Module S.Subrange

type t

The type of subranges. Each subrange is a contiguous region of code delimited by labels.

val info : t -> Subrange_info.t

The caller's information about the subrange.

val start_pos : t -> Linear.label

The label at the start of the range.

val start_pos_offset : t -> int

How many bytes from the label at start_pos the range actually commences. If this value is zero, then the first byte of the range has the address of the label given by start_pos.

val end_pos : t -> Linear.label

The label at the end of the range.

val end_pos_offset : t -> int

Like start_pos_offset, but analogously for the end of the range. (The sense is not inverted; a positive end_pos_offset means the range ends at an address higher than the address of the end_pos, just like a positive start_pos_offset means the range starts at an address higher than the start_pos.