Module Define.Simple

Parameters

module _ : Parameters with type update_result := D.state

Signature

val diff : D.state -> D.left array -> D.right array -> patch

diff state l r computes the optimal patch between l and r, using the initial state state.