Module Docstrings.WithMenhir

val symbol_docs : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> docs

Fetch the item documentation for the current symbol. This also marks this documentation (for ambiguity warnings).

val symbol_docs_lazy : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> docs Stdlib.Lazy.t

val rhs_docs : Stdlib.Lexing.position -> Stdlib.Lexing.position -> docs

Fetch the item documentation for the symbols between two positions. This also marks this documentation (for ambiguity warnings).

val rhs_docs_lazy : Stdlib.Lexing.position -> Stdlib.Lexing.position -> docs Stdlib.Lazy.t

val mark_symbol_docs : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> unit

Mark the item documentation for the current symbol (for ambiguity warnings).

val mark_rhs_docs : Stdlib.Lexing.position -> Stdlib.Lexing.position -> unit

Mark as associated the item documentation for the symbols between two positions (for ambiguity warnings)

val symbol_info : Stdlib.Lexing.position -> info

Fetch the field info for the current symbol.

val rhs_info : Stdlib.Lexing.position -> info

Fetch the field info following the symbol at a given position.

val symbol_text : Stdlib.Lexing.position -> text

Fetch the text preceding the current symbol.

val symbol_text_lazy : Stdlib.Lexing.position -> text Stdlib.Lazy.t

val rhs_text : Stdlib.Lexing.position -> text

Fetch the text preceding the symbol at the given position.

val rhs_text_lazy : Stdlib.Lexing.position -> text Stdlib.Lazy.t

Extra text

There may be additional text attached to the delimiters of a block (e.g. struct and end). This is fetched by the following functions, which are applied to the contents of the block rather than the delimiters.

val symbol_pre_extra_text : Stdlib.Lexing.position -> text

Fetch additional text preceding the current symbol

val symbol_post_extra_text : Stdlib.Lexing.position -> text

Fetch additional text following the current symbol

val rhs_pre_extra_text : Stdlib.Lexing.position -> text

Fetch additional text preceding the symbol at the given position

val rhs_post_extra_text : Stdlib.Lexing.position -> text

Fetch additional text following the symbol at the given position

val rhs_post_text : Stdlib.Lexing.position -> text

Fetch text following the symbol at the given position