Module Export_info_for_pack

Transformations on export information that are only used for the building of packs.

val import_for_pack : pack_units:Compilation_unit.Set.t -> pack:Compilation_unit.t -> Export_info.t -> Export_info.t

Transform the information from exported to be suitable to be reexported as the information for a pack named pack containing units pack_units. It mainly changes symbols of units pack_units to refer to pack instead.

val clear_import_state : unit -> unit

Drops the state after importing several units in the same pack.