Module Find_recursive_functions

"Recursive functions" are those functions f that might call either:

themselves, or

another function that in turn might call f.

For example in the following simultaneous definition of f g and h, f and g are recursive functions, but not h: let rec f x = g x and g x = f x and h x = g x

val in_function_declarations : Flambda.function_declarations -> backend:(module Backend_intf.S) -> Variable.Set.t

Determine the recursive functions, if any, bound by the given set of function declarations. This is only intended to be used by Flambda.create_function_declarations.