Module type Identifiable.S

type t

module T : Thing with type t = t

include Thing with type t := T.t

include Stdlib.Hashtbl.HashedType with type t := T.t

val equal : T.t -> T.t -> bool

The equality predicate used to compare keys.

val hash : T.t -> int

A hashing function on keys. It must be such that if two keys are equal according to equal, then they have identical hash values as computed by hash. Examples: suitable (equal, hash) pairs for arbitrary key types include

((=), Hashtbl.hash) for comparing objects by structure (provided objects do not contain floats)

((fun x y -> compare x y = 0), Hashtbl.hash) for comparing objects by structure and handling Stdlib.nan correctly

((==), Hashtbl.hash) for comparing objects by physical equality (e.g. for mutable or cyclic objects).

include Stdlib.Map.OrderedType with type t := T.t

val compare : T.t -> T.t -> int

A total ordering function over the keys. This is a two-argument function f such that f e1 e2 is zero if the keys e1 and e2 are equal, f e1 e2 is strictly negative if e1 is smaller than e2, and f e1 e2 is strictly positive if e1 is greater than e2. Example: a suitable ordering function is the generic structural comparison function Stdlib.compare.

val output : Stdlib.out_channel -> T.t -> unit

val print : Stdlib.Format.formatter -> T.t -> unit

module Set : Set with module T := T

module Map : Map with module T := T

module Tbl : Tbl with module T := T