Module Main_args

module type Common_options = sig ... end

module type Core_options = sig ... end

module type Compiler_options = sig ... end

module type Toplevel_options = sig ... end

module type Bytecomp_options = sig ... end

module type Bytetop_options = sig ... end

module type Optcommon_options = sig ... end

module type Optcomp_options = sig ... end

module type Opttop_options = sig ... end

module type Ocamldoc_options = sig ... end

module type Arg_list = sig ... end

module Make_bytecomp_options (_ : Bytecomp_options) : Arg_list

module Make_bytetop_options (_ : Bytetop_options) : Arg_list

module Make_optcomp_options (_ : Optcomp_options) : Arg_list

module Make_opttop_options (_ : Opttop_options) : Arg_list

module Make_ocamldoc_options (_ : Ocamldoc_options) : Arg_list

val options_with_command_line_syntax : (string * Stdlib.Arg.spec * string) list -> string list ref -> (string * Stdlib.Arg.spec * string) list

options_with_command_line_syntax options r returns options2 that behaves like options, but additionally pushes command line argument on r (quoted by Filename.quote when necessary). This is meant for ocamlc,optp, which use this to forward most of their arguments to ocamlc,opt.

module Default : sig ... end