Module Stdlib.Array

val exists2 : ('a -> 'b -> bool) -> 'a array -> 'b array -> bool

val for_alli : (int -> 'a -> bool) -> 'a array -> bool

Same as Array.for_all, but the function is applied with the index of the element as first argument, and the element itself as second argument.

val all_somes : 'a option array -> 'a array option