Module Odoc_info.Dep

Computation of dependencies.

val kernel_deps_of_modules : Module.t_module list -> unit

Modify the module dependencies of the given list of modules, to get the minimum transitivity kernel.

val deps_of_types : ?kernel:bool -> Type.t_type list -> (Type.t_type * Name.t list) list

Return the list of dependencies between the given types, in the form of a list (type name, names of types it depends on).

parameter kernel
indicates if we must keep only the transitivity kernel of the dependencies. Default is false.