Module Odoc_info.Global

val errors : int ref

val warn_error : bool ref

val out_file : string ref

The file used by the generators outputting only one file.

val verbose : bool ref

Verbose mode or not.

val target_dir : string ref

The directory where files have to be generated.

val title : string option ref

The optional title to use in the generated documentation.

val intro_file : string option ref

The optional file whose content can be used as intro text.

val with_toc : bool ref

The flag which indicates if we must generate a table of contents.

val with_index : bool ref

The flag which indicates if we must generate an index.

val with_header : bool ref

The flag which indicates if we must generate a header.

val with_trailer : bool ref

The flag which indicates if we must generate a trailer.