Class Scan.scanner

method scan_value : Value.t_value -> unit

method scan_type_pre : Type.t_type -> bool

method scan_type_const : Type.t_type -> Type.variant_constructor -> unit

method scan_type_recfield : Type.t_type -> Type.record_field -> unit

method scan_type : Type.t_type -> unit

method scan_extension_constructor : Extension.t_extension_constructor -> unit

method scan_exception : Exception.t_exception -> unit

method scan_attribute : Value.t_attribute -> unit

method scan_method : Value.t_method -> unit

method scan_included_module : Module.included_module -> unit

Scan of a type extension

method scan_type_extension_pre : Extension.t_type_extension -> bool

Override this method to perform controls on the extension's type, private and info. This method is called before scanning the extension's constructors.

returns
true if the extension's constructors must be scanned.

method scan_type_extension_constructors : Extension.t_type_extension -> unit

This method scans the constructors of the given type extension.

method scan_type_extension : Extension.t_type_extension -> unit

Scan of a type extension. Should not be overridden. It calls scan_type_extension_pre and if scan_type_extension_pre returns true, then it calls scan_type_extension_constructors.

Scan of a class.

method scan_class_comment : text -> unit

Scan of a comment inside a class.

method scan_class_pre : Class.t_class -> bool

Override this method to perform controls on the class comment and params. This method is called before scanning the class elements.

returns
true if the class elements must be scanned.

method scan_class_elements : Class.t_class -> unit

This method scans the elements of the given class.

method scan_class : Class.t_class -> unit

Scan of a class. Should not be overridden. It calls scan_class_pre and if scan_class_pre returns true, then it calls scan_class_elements.

Scan of a class type.

method scan_class_type_comment : text -> unit

Scan of a comment inside a class type.

method scan_class_type_pre : Class.t_class_type -> bool

Override this method to perform controls on the class type comment and form. This method is called before scanning the class type elements.

returns
true if the class type elements must be scanned.

method scan_class_type_elements : Class.t_class_type -> unit

This method scans the elements of the given class type.

method scan_class_type : Class.t_class_type -> unit

Scan of a class type. Should not be overridden. It calls scan_class_type_pre and if scan_class_type_pre returns true, then it calls scan_class_type_elements.

Scan of modules.

method scan_module_comment : text -> unit

Scan of a comment inside a module.

method scan_module_pre : Module.t_module -> bool

Override this method to perform controls on the module comment and form. This method is called before scanning the module elements.

returns
true if the module elements must be scanned.

method scan_module_elements : Module.t_module -> unit

This method scans the elements of the given module.

method scan_module : Module.t_module -> unit

Scan of a module. Should not be overridden. It calls scan_module_pre and if scan_module_pre returns true, then it calls scan_module_elements.

Scan of module types.

method scan_module_type_comment : text -> unit

Scan of a comment inside a module type.

method scan_module_type_pre : Module.t_module_type -> bool

Override this method to perform controls on the module type comment and form. This method is called before scanning the module type elements.

returns
true if the module type elements must be scanned.

method scan_module_type_elements : Module.t_module_type -> unit

This method scans the elements of the given module type.

method scan_module_type : Module.t_module_type -> unit

Scan of a module type. Should not be overridden. It calls scan_module_type_pre and if scan_module_type_pre returns true, then it calls scan_module_type_elements.

Main scanning method.

method scan_module_list : Module.t_module list -> unit

Scan a list of modules.