Module Odoc_info.Value

Representation and manipulation of values, class attributes and class methods.

type t_value = Odoc_value.t_value = {

val_name : Name.t;
Complete name of the value.

mutable val_info : info option;
Information found in the optional associated comment.

val_type : Types.type_expr;
Type of the value.

val_recursive : bool;
true if the value is recursive.

mutable val_parameters : Odoc_parameter.parameter list;
The parameters, if any.

mutable val_code : string option;
The code of the value, if we had the only the implementation file.

mutable val_loc : location;

}

Representation of a value.

type t_attribute = Odoc_value.t_attribute = {

att_value : t_value;
an attribute has almost all the same information as a value

att_mutable : bool;
true if the attribute is mutable.

att_virtual : bool;
true if the attribute is virtual.

}

Representation of a class attribute.

type t_method = Odoc_value.t_method = {

met_value : t_value;
a method has almost all the same information as a value

met_private : bool;
true if the method is private.

met_virtual : bool;
true if the method is virtual.

}

Representation of a class method.

val is_function : t_value -> bool

Return true if the value is a function, i.e. it has a functional type.

val value_parameter_text_by_name : t_value -> string -> text option

Access to the description associated to the given parameter name.