Module Out_type.Variable_names

Naming choice for type variable names ('a, 'b, ...), for instance the two classes of distinct type variables in

let repeat x y = x, y, y, x

should be printed printed as 'a -> 'b -> 'a * 'b * 'b * 'a.

val add_subst : (Types.type_expr * Types.type_expr) list -> unit

Add external type equalities

val reserve : Types.type_expr -> unit

reserve ty registers the variable names appearing in ty