Module Pprintast

Pretty-printers for Parsetree

Warning: this module is unstable and part of compiler-libs.

type space_formatter = (unit, Stdlib.Format.formatter, unit) Stdlib.format

val longident : Stdlib.Format.formatter -> Longident.t -> unit

val expression : Stdlib.Format.formatter -> Parsetree.expression -> unit

val string_of_expression : Parsetree.expression -> string

val pattern : Stdlib.Format.formatter -> Parsetree.pattern -> unit

val core_type : Stdlib.Format.formatter -> Parsetree.core_type -> unit

val signature : Stdlib.Format.formatter -> Parsetree.signature -> unit

val structure : Stdlib.Format.formatter -> Parsetree.structure -> unit

val string_of_structure : Parsetree.structure -> string

val toplevel_phrase : Stdlib.Format.formatter -> Parsetree.toplevel_phrase -> unit

val top_phrase : Stdlib.Format.formatter -> Parsetree.toplevel_phrase -> unit

val tyvar : Stdlib.Format.formatter -> string -> unit

Print a type variable name, taking care of the special treatment required for the single quote character in second position.