Module Pprintast.Doc

Format_doc functions for error messages

val longident : Longident.t Format_doc.printer

val constr : Longident.t Format_doc.printer

val tyvar : string Format_doc.printer

val nominal_exp : Parsetree.expression -> Format_doc.t option

Returns a format document if the expression reads nicely as the subject of a sentence in a error message.