Module Raw_spacetime_lib

Access to the information recorded by the Spacetime module. (It is intended that this module will be used by post-processors rather than users wishing to understand their programs.) For 64-bit targets only. This module may be used from any program, not just one compiled with a compiler configured for Spacetime.

module Gc_stats : sig ... end

module Annotation : sig ... end

module Program_counter : sig ... end

module Frame_table : sig ... end

module Function_entry_point : sig ... end

module Function_identifier : sig ... end

module Shape_table : sig ... end

module Trace : sig ... end

module Heap_snapshot : sig ... end