Module Raw_spacetime_lib.Frame_table

type t

A value of type t corresponds to the frame table of a running OCaml program. The table is indexed by program counter address (typically, but not always when using Spacetime, return addresses).

val find_exn : Program_counter.OCaml.t -> t -> Stdlib.Printexc.Slot.t list

Find the location, including any inlined frames, corresponding to the given program counter address. Raises Not_found if the location could not be resolved.