Module Heap_snapshot.Entries

type t

An immutable array of the total number of blocks (= boxed values) and the total number of words occupied by such blocks (including their headers) for each profiling annotation in the heap.

val length : t -> int

val annotation : t -> int -> Annotation.t

val num_blocks : t -> int -> int

val num_words_including_headers : t -> int -> int