Module Raw_spacetime_lib.Trace

type t

A value of type t holds the dynamic call structure of the program (i.e. which functions have called which other functions) together with information required to decode profiling annotations written into values' headers.

type trace = t

type node

type ocaml_node

type foreign_node

type uninstrumented_node

module OCaml : sig ... end

module Foreign : sig ... end

module Node : sig ... end

val root : t -> Node.t option

Obtains the root of the graph for traversal. None is returned if the graph is empty.