Module OCaml.Allocation_point

type t

A value of type t corresponds to an allocation point in OCaml code.

val program_counter : t -> Program_counter.OCaml.t

The program counter at (or close to) the allocation site.

val annotation : t -> Annotation.t

The annotation written into the headers of boxed values allocated at the given allocation site.

val num_words_including_headers : t -> int

The total number of words allocated at this point.