Module OCaml.Direct_call_point

type 'target t

A value of type 'target t corresponds to a direct (i.e. known at compile time) call point in OCaml code. 'target is the type of the node corresponding to the callee.

val call_site : _ t -> Program_counter.OCaml.t

The program counter at (or close to) the call site.

val callee : _ t -> Function_entry_point.t

The address of the first instruction of the callee.

val callee_node : 'target t -> 'target

The node corresponding to the callee.

val call_count : _ t -> int option

The number of times the callee was called. Only available if the compiler that recorded the Spacetime profile was configured with "-with-spacetime-call-counts". None will be returned otherwise.