Module OCaml.Indirect_call_point

type t

A value of type t corresponds to an indirect call point in OCaml code. Each such value contains a list of callees to which the call point has branched.

val call_site : t -> Program_counter.OCaml.t

The program counter at (or close to) the call site.

module Callee : sig ... end

val callees : t -> Callee.t option

The list of callees to which this indirect call point has branched.