Module Shape

Shapes are an abstract representation of modules' implementations which allow the tracking of definitions through functor applications and other module-level operations.

The Shape of a compilation unit is elaborated during typing, partially reduced (without loading external shapes) and written to the cmt file.

External tools can retrieve the definition of any value (or type, or module, etc) by following this procedure:

Build the Shape corresponding to the value's path: let shape = Env.shape_of_path ~namespace env path

Instantiate the Shape_reduce.Make functor with a way to load shapes from external units and to looks for shapes in the environment (usually using Env.shape_of_path).

Completely reduce the shape: let shape = My_reduce.(weak_)reduce env shape

The Uid.t stored in the reduced shape should be the one of the definition. However, if the approximate field of the reduced shape is true then the Uid.t will not correspond to the definition, but to the closest parent module's uid. This happens when Shape reduction gets stuck, for example when hitting first-class modules.

The location of the definition can be easily found with the cmt_format.cmt_uid_to_decl table of the corresponding compilation unit.

See:

the design document

a talk about the reduction strategy

module Uid : sig ... end

A Uid.t is associated to every declaration in signatures and implementations. They uniquely identify bindings in the program. When associated with these bindings' locations they are useful to external tools when trying to jump to an identifier's declaration or definition. They are stored to that effect in the uid_to_decl table of cmt files.

module Sig_component_kind : sig ... end

module Item : sig ... end

Shape's items are elements of a structure or, in the case of constructors and labels, elements of a record or variants definition seen as a structure. These structures model module components and nested types' constructors and labels.

type var = Ident.t

type t = {

uid : Uid.t option;

desc : desc;

approximated : bool;

}

and desc =

| Var of var

| Abs of var * t

| App of t * t

| Struct of t Item.Map.t

| Alias of t

| Leaf

| Proj of t * Item.t

| Comp_unit of string

| Error of string

val print : Stdlib.Format.formatter -> t -> unit

val strip_head_aliases : t -> t

val for_unnamed_functor_param : var

val fresh_var : ?name:string -> Uid.t -> var * t

val var : Uid.t -> Ident.t -> t

val abs : ?uid:Uid.t -> var -> t -> t

val app : ?uid:Uid.t -> t -> arg:t -> t

val str : ?uid:Uid.t -> t Item.Map.t -> t

val alias : ?uid:Uid.t -> t -> t

val proj : ?uid:Uid.t -> t -> Item.t -> t

val leaf : Uid.t -> t

val decompose_abs : t -> (var * t) option

val for_persistent_unit : string -> t

val leaf_for_unpack : t

module Map : sig ... end

val dummy_mod : t

val of_path : find_shape:(Sig_component_kind.t -> Ident.t -> t) -> namespace:Sig_component_kind.t -> Path.t -> t

This function returns the shape corresponding to a given path. It requires a callback to find shapes in the environment. It is generally more useful to rely directly on the Env.shape_of_path function to get the shape associated with a given path.

val set_uid_if_none : t -> Uid.t -> t