Module Stdlib.Effect

alert unstable The Effect interface may change in incompatible ways in the future.

Effects.

See 'Language extensions/Effect handlers' section in the manual.

alert unstable The Effect interface may change in incompatible ways in the future.

type 'a t = 'a eff = ..

The type of effects.

exception Unhandled : 'a t -> exn

Unhandled e is raised when effect e is performed and there is no handler for it.

exception Continuation_already_resumed

Exception raised when a continuation is continued or discontinued more than once.

val perform : 'a t -> 'a

perform e performs an effect e.

raises Unhandled
if there is no handler for e.

module Deep : sig ... end

Deep handlers

module Shallow : sig ... end