Module Ephemeron.GenHashTable

Hash tables on generic containers with notion of death and aliveness.

Define a hash table on generic containers which have a notion of "death" and aliveness. If a binding is dead the hash table can automatically remove it.

alert old_ephemeron_api This module won't be available in 5.0

type equal =

| ETrue

| EFalse

| EDead
(*
the container is dead
*)

module MakeSeeded (H : sig ... end) : SeededS with type key = H.t

Functor building an implementation of an hash table that use the container for keeping the information given