Module Ephemeron.K2

type ('k1, 'k2, 'd) t

an ephemeron with two keys

val create : unit -> ('k1, 'k2, 'd) t

Same as Ephemeron.K1.create

val get_key1 : ('k1, 'k2, 'd) t -> 'k1 option

Same as Ephemeron.K1.get_key

val get_key1_copy : ('k1, 'k2, 'd) t -> 'k1 option

Same as Ephemeron.K1.get_key_copy

val set_key1 : ('k1, 'k2, 'd) t -> 'k1 -> unit

Same as Ephemeron.K1.set_key

val unset_key1 : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_key

val check_key1 : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_key

val get_key2 : ('k1, 'k2, 'd) t -> 'k2 option

Same as Ephemeron.K1.get_key

val get_key2_copy : ('k1, 'k2, 'd) t -> 'k2 option

Same as Ephemeron.K1.get_key_copy

val set_key2 : ('k1, 'k2, 'd) t -> 'k2 -> unit

Same as Ephemeron.K1.set_key

val unset_key2 : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_key

val check_key2 : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_key

val blit_key1 : ('k1, _, _) t -> ('k1, _, _) t -> unit

Same as Ephemeron.K1.blit_key

val blit_key2 : (_, 'k2, _) t -> (_, 'k2, _) t -> unit

Same as Ephemeron.K1.blit_key

val blit_key12 : ('k1, 'k2, _) t -> ('k1, 'k2, _) t -> unit

Same as Ephemeron.K1.blit_key

val get_data : ('k1, 'k2, 'd) t -> 'd option

Same as Ephemeron.K1.get_data

val get_data_copy : ('k1, 'k2, 'd) t -> 'd option

Same as Ephemeron.K1.get_data_copy

val set_data : ('k1, 'k2, 'd) t -> 'd -> unit

Same as Ephemeron.K1.set_data

val unset_data : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_data

val check_data : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_data

val blit_data : ('k1, 'k2, 'd) t -> ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.blit_data

module Make : functor (H1 : Stdlib.Hashtbl.HashedType) -> functor (H2 : Stdlib.Hashtbl.HashedType) -> S with type key = H1.t * H2.t

Functor building an implementation of a weak hash table

module MakeSeeded : functor (H1 : Stdlib.Hashtbl.SeededHashedType) -> functor (H2 : Stdlib.Hashtbl.SeededHashedType) -> SeededS with type key = H1.t * H2.t

Functor building an implementation of a weak hash table. The seed is similar to the one of Hashtbl.MakeSeeded.