Module Ephemeron.Kn

Ephemerons with arbitrary number of keys of the same type.

type ('k, 'd) t

an ephemeron with an arbitrary number of keys of the same type

val create : int -> ('k, 'd) t

Same as Ephemeron.K1.create

alert old_ephemeron_api This function won't be available in 5.0

val get_key : ('k, 'd) t -> int -> 'k option

Same as Ephemeron.K1.get_key

alert old_ephemeron_api This function won't be available in 5.0

val get_key_copy : ('k, 'd) t -> int -> 'k option

Same as Ephemeron.K1.get_key_copy

alert old_ephemeron_api This function won't be available in 5.0

val set_key : ('k, 'd) t -> int -> 'k -> unit

Same as Ephemeron.K1.set_key

alert old_ephemeron_api This function won't be available in 5.0

val unset_key : ('k, 'd) t -> int -> unit

Same as Ephemeron.K1.unset_key

alert old_ephemeron_api This function won't be available in 5.0

val check_key : ('k, 'd) t -> int -> bool

Same as Ephemeron.K1.check_key

alert old_ephemeron_api This function won't be available in 5.0

val blit_key : ('k, _) t -> int -> ('k, _) t -> int -> int -> unit

Same as Ephemeron.K1.blit_key

alert old_ephemeron_api This function won't be available in 5.0

val get_data : ('k, 'd) t -> 'd option

Same as Ephemeron.K1.get_data

alert old_ephemeron_api This function won't be available in 5.0

val get_data_copy : ('k, 'd) t -> 'd option

Same as Ephemeron.K1.get_data_copy

alert old_ephemeron_api This function won't be available in 5.0

val set_data : ('k, 'd) t -> 'd -> unit

Same as Ephemeron.K1.set_data

alert old_ephemeron_api This function won't be available in 5.0

val unset_data : ('k, 'd) t -> unit

Same as Ephemeron.K1.unset_data

alert old_ephemeron_api This function won't be available in 5.0

val check_data : ('k, 'd) t -> bool

Same as Ephemeron.K1.check_data

alert old_ephemeron_api This function won't be available in 5.0

val blit_data : ('k, 'd) t -> ('k, 'd) t -> unit

Same as Ephemeron.K1.blit_data

alert old_ephemeron_api This function won't be available in 5.0

val make : 'k array -> 'd -> ('k, 'd) t

Same as Ephemeron.K1.make

val query : ('k, 'd) t -> 'k array -> 'd option

Same as Ephemeron.K1.query

module Make (H : Hashtbl.HashedType) : S with type key = H.t array

Functor building an implementation of a weak hash table

module MakeSeeded (H : Hashtbl.SeededHashedType) : SeededS with type key = H.t array

Functor building an implementation of a weak hash table. The seed is similar to the one of Hashtbl.MakeSeeded.

module Bucket : sig ... end