Module Kn.Make

Functor building an implementation of a weak hash table

Propose the same interface as usual hash table. However since the bindings are weak, even if mem h k is true, a subsequent find h k may raise Not_found because the garbage collector can run between the two.

Parameters

module H : Hashtbl.HashedType

Signature

type key = H.t array

type !'a t

val create : int -> 'a t

val clear : 'a t -> unit

val reset : 'a t -> unit

val copy : 'a t -> 'a t

val add : 'a t -> key -> 'a -> unit

val remove : 'a t -> key -> unit

val find : 'a t -> key -> 'a

val find_opt : 'a t -> key -> 'a option

val find_all : 'a t -> key -> 'a list

val replace : 'a t -> key -> 'a -> unit

val mem : 'a t -> key -> bool

val length : 'a t -> int

val stats : 'a t -> Hashtbl.statistics

val add_seq : 'a t -> (key * 'a) Seq.t -> unit

val replace_seq : 'a t -> (key * 'a) Seq.t -> unit

val of_seq : (key * 'a) Seq.t -> 'a t

val clean : 'a t -> unit

remove all dead bindings. Done automatically during automatic resizing.

val stats_alive : 'a t -> Hashtbl.statistics

same as Hashtbl.SeededS.stats but only count the alive bindings