Module Stdlib.Sys

val argv : string array

val executable_name : string

val file_exists : string -> bool

val is_directory : string -> bool

val is_regular_file : string -> bool

val remove : string -> unit

val rename : string -> string -> unit

val getenv : string -> string

val getenv_opt : string -> string option

val command : string -> int

val time : unit -> float

val chdir : string -> unit

val mkdir : string -> int -> unit

val rmdir : string -> unit

val getcwd : unit -> string

val readdir : string -> string array

val interactive : bool ref

alert unsynchronized_access The interactive status is a mutable global state.

val os_type : string

type backend_type =

| Native

| Bytecode

| Other of string

val backend_type : backend_type

val unix : bool

val win32 : bool

val cygwin : bool

val word_size : int

val int_size : int

val big_endian : bool

val max_string_length : int

val max_array_length : int

val max_floatarray_length : int

val runtime_variant : unit -> string

val runtime_parameters : unit -> string

val poll_actions : unit -> unit

type signal_behavior =

| Signal_default

| Signal_ignore

| Signal_handle of int -> unit

val signal : int -> signal_behavior -> signal_behavior

val set_signal : int -> signal_behavior -> unit

val sigabrt : int

val sigalrm : int

val sigfpe : int

val sighup : int

val sigill : int

val sigint : int

val sigkill : int

val sigpipe : int

val sigquit : int

val sigsegv : int

val sigterm : int

val sigusr1 : int

val sigusr2 : int

val sigchld : int

val sigcont : int

val sigstop : int

val sigtstp : int

val sigttin : int

val sigttou : int

val sigvtalrm : int

val sigprof : int

val sigbus : int

val sigpoll : int

val sigsys : int

val sigtrap : int

val sigurg : int

val sigxcpu : int

val sigxfsz : int

exception Break

val catch_break : bool -> unit

val ocaml_version : string

val development_version : bool

type extra_prefix =

| Plus

| Tilde

type extra_info = extra_prefix * string

type ocaml_release_info = {

major : int;

minor : int;

patchlevel : int;

extra : extra_info option;

}

val ocaml_release : ocaml_release_info

val enable_runtime_warnings : bool -> unit

alert unsynchronized_access The status of runtime warnings is a mutable global state.

val runtime_warnings_enabled : unit -> bool

alert unsynchronized_access The status of runtime warnings is a mutable global state.

val opaque_identity : 'a -> 'a

module Immediate64 : sig ... end