Module Stdlib__arg

Parsing of command line arguments.

This module provides a general mechanism for extracting options and arguments from the command line to the program.

Syntax of command lines: A keyword is a character string starting with a -. An option is a keyword alone or followed by an argument. The types of keywords are: Unit, Bool, Set, Clear, String, Set_string, Int, Set_int, Float, Set_float, Tuple, Symbol, and Rest. Unit, Set and Clear keywords take no argument. A Rest keyword takes the remaining of the command line as arguments. Every other keyword takes the following word on the command line as argument. For compatibility with GNU getopt_long, keyword=arg is also allowed. Arguments not preceded by a keyword are called anonymous arguments.

Examples (cmd is assumed to be the command name):

cmd -flag (a unit option)

cmd -int 1 (an int option with argument 1)

cmd -string foobar (a string option with argument "foobar")

cmd -float 12.34 (a float option with argument 12.34)

cmd a b c (three anonymous arguments: "a", "b", and "c")

cmd a b -- c d (two anonymous arguments and a rest option with two arguments)