Module Stdlib__arrayLabels

Array operations

This module is intended to be used via StdLabels which replaces Array, Bytes, List and String with their labeled counterparts

For example:

open StdLabels

let everything = Array.create_matrix ~dimx:42 ~dimy:42 42