Module Stdlib__bytesLabels

Byte sequence operations.

since

4.02.0

This module is intended to be used through StdLabels which replaces Array, Bytes, List and String with their labeled counterparts.

For example:

open StdLabels

let first = Bytes.sub ~pos:0 ~len:1