Module Stdlib__moreLabels

Extra labeled libraries.

This meta-module provides labelized version of the Hashtbl, Map and Set modules.

They only differ by their labels. They are provided to help porting from previous versions of OCaml. The contents of this module are subject to change.