Module Stdlib__stringLabels

String operations. This module is intended to be used through StdLabels which replaces Array, Bytes, List and String with their labeled counterparts

For example:

open StdLabels

let to_upper = String.map ~f:Char.uppercase_ascii