Module Tag

Tags on runtime boxed values.

include Identifiable.S

type t

module T : Identifiable.Thing with type t = t

include Identifiable.Thing with type t := T.t

include Hashtbl.HashedType with type t := T.t

val equal : T.t -> T.t -> bool

The equality predicate used to compare keys.

val hash : T.t -> int

A hashing function on keys. It must be such that if two keys are equal according to equal, then they have identical hash values as computed by hash. Examples: suitable (equal, hash) pairs for arbitrary key types include

((=), hash) for comparing objects by structure (provided objects do not contain floats)

((fun x y -> compare x y = 0), hash) for comparing objects by structure and handling Stdlib.nan correctly

((==), hash) for comparing objects by physical equality (e.g. for mutable or cyclic objects).

include Map.OrderedType with type t := T.t

val output : out_channel -> T.t -> unit

val print : Stdlib.Format.formatter -> T.t -> unit

module Set : Identifiable.Set with module T := T

module Map : Identifiable.Map with module T := T

module Tbl : Identifiable.Tbl with module T := T

val create_exn : int -> t

val to_int : t -> int

val zero : t

val object_tag : t

val compare : t -> t -> int