Module Topcommon

This module provides common implementations for internals of Toploop, for bytecode and native code (see Topeval for the diverging parts of the implementation).

You should not use it directly, refer to the functions in Toploop instead.

type evaluation_outcome =

| Result of Stdlib.Obj.t

| Exception of exn

val backtrace : string option ref

val parse_mod_use_file : string -> Stdlib.Lexing.lexbuf -> Parsetree.toplevel_phrase list

val comment_prompt_override : bool ref

val refill_lexbuf : bytes -> int -> int