Module Unit_info.Artifact

Build artifacts

type t

Metadata for a single compilation artifact:

the module name associated to the artifact

the filesystem path

the input source file if it exists

val source_file : t -> filename option

source_file a is the source file of a if it exists.

val prefix : t -> file_prefix

prefix a is the filename prefix of the compilation artifact.

val filename : t -> filename

filename u is the filesystem path for a compilation artifact.

val modname : t -> modname

modname a is the module name of the compilation artifact.

val from_filename : filename -> t

from_filename filename reconstructs the module name modname_from_source filename associated to the artifact filename.