Module Variable

Variable.t is the equivalent of a non-persistent Ident.t in the Flambda tree. It wraps an Ident.t together with its source compilation_unit. As such, it is unique within a whole program, not just one compilation unit.

Introducing a new type helps in tracing the source of identifiers when debugging the inliner. It also avoids Ident renaming when importing cmx files.

include Identifiable.S

type t

module T : Identifiable.Thing with type t = t

include Identifiable.Thing with type t := T.t

include Hashtbl.HashedType with type t := T.t

val equal : T.t -> T.t -> bool

The equality predicate used to compare keys.

val hash : T.t -> int

A hashing function on keys. It must be such that if two keys are equal according to equal, then they have identical hash values as computed by hash. Examples: suitable (equal, hash) pairs for arbitrary key types include

((=), hash) for comparing objects by structure (provided objects do not contain floats)

((fun x y -> compare x y = 0), hash) for comparing objects by structure and handling Stdlib.nan correctly

((==), hash) for comparing objects by physical equality (e.g. for mutable or cyclic objects).

include Map.OrderedType with type t := T.t

val compare : T.t -> T.t -> int

A total ordering function over the keys. This is a two-argument function f such that f e1 e2 is zero if the keys e1 and e2 are equal, f e1 e2 is strictly negative if e1 is smaller than e2, and f e1 e2 is strictly positive if e1 is greater than e2. Example: a suitable ordering function is the generic structural comparison function Stdlib.compare.

val output : out_channel -> T.t -> unit

val print : Stdlib.Format.formatter -> T.t -> unit

module Set : Identifiable.Set with module T := T

module Map : Identifiable.Map with module T := T

module Tbl : Identifiable.Tbl with module T := T

val create : ?current_compilation_unit:Compilation_unit.t -> Internal_variable_names.t -> t

val create_with_same_name_as_ident : Ident.t -> t

val rename : ?current_compilation_unit:Compilation_unit.t -> t -> t

val in_compilation_unit : t -> Compilation_unit.t -> bool

val name : t -> string

val unique_name : t -> string

val get_compilation_unit : t -> Compilation_unit.t

val print_list : Stdlib.Format.formatter -> t list -> unit

val print_opt : Stdlib.Format.formatter -> t option -> unit

val debug_when_stamp_matches : t -> stamp:int -> f:(unit -> unit) -> unit

If the given variable has the given stamp, call the user-supplied function. For debugging purposes only.

type pair = t * t

module Pair : Identifiable.S with type t := pair

val compare_lists : t list -> t list -> int

val output_full : out_channel -> t -> unit

Unlike output, output_full includes the compilation unit.