Module Attr_helper

Helpers for attributes

Warning: this module is unstable and part of compiler-libs.

type error =

| Multiple_attributes of string

| No_payload_expected of string

val get_no_payload_attribute : string -> Parsetree.attributes -> string Asttypes.loc option

The string argument of the following functions is the name of the attribute we are looking for. If the argument is "foo", these functions will find attributes with the name "foo" or "ocaml.foo"

val has_no_payload_attribute : string -> Parsetree.attributes -> bool

exception Error of Location.t * error

val report_error : error Format_doc.format_printer

val report_error_doc : error Format_doc.printer