Module 1-S.Range_info

The module Range_info is used to store additional information on a range that is associated to a range at its creation and can be retrieved from the result. The association between keys and indices is also done here: Range_info.create serves both as a map between keys and indices; and also as the creator of the Range_info.t structure. When several subranges are contained in a single range, the associated Range_info.t will correspond to the first closed subrange.

type t

val create : Compute_ranges_intf.L.fundecl -> Key.t -> start_insn:Compute_ranges_intf.L.instruction -> (Index.t * t) option