Module Import_approx

Create simple value approximations from the export information in .cmx files.

val really_import : Simple_value_approx.descr -> Simple_value_approx.descr

Given an approximation description, load .cmx files (possibly more than one) until the description is fully resolved. If a necessary .cmx file cannot be found, "unresolved" will be returned.

val really_import_approx : Simple_value_approx.t -> Simple_value_approx.t

Maps the description of the given approximation through really_import.

val import_symbol : Symbol.t -> Simple_value_approx.t

Read and convert the approximation of a given symbol from the relevant .cmx file. Unlike the "really_" functions, this does not continue to load .cmx files until the approximation is fully resolved.