Module Inlining_cost

Measurement of the cost (including cost in space) of Flambda terms in the context of inlining.

module Threshold : sig ... end

val can_inline : Flambda.t -> Threshold.t -> bonus:int -> bool

val can_try_inlining : Flambda.t -> Threshold.t -> number_of_arguments:int -> size_from_approximation:int option -> Threshold.t

module Benefit : sig ... end

module Whether_sufficient_benefit : sig ... end

val scale_inline_threshold_by : int

val default_toplevel_multiplier : int

val direct_call_size : int

val maximum_interesting_size_of_function_body : int -> int

If a function body exceeds this size, we can make a fast decision not to inline it (see Inlining_decision).

val lambda_smaller' : Flambda.expr -> than:int -> int option

Measure the given expression to determine whether its size is at or below the given threshold. None is returned if it is too big; otherwise Some is returned with the measured size.

val lambda_size : Flambda.expr -> int