Module Inlining_cost.Threshold

type t =

| Never_inline

| Can_inline_if_no_larger_than of int

The maximum size, in some abstract measure of space cost, that an Flambda expression may be in order to be inlined.

val add : t -> t -> t

val sub : t -> t -> t

val min : t -> t -> t

val equal : t -> t -> bool