Module Odoc_info.Search

Research in elements

type result_element = Odoc_search.result_element =

| Res_module of Module.t_module

| Res_module_type of Module.t_module_type

| Res_class of Class.t_class

| Res_class_type of Class.t_class_type

| Res_value of Value.t_value

| Res_type of Type.t_type

| Res_extension of Extension.t_extension_constructor

| Res_exception of Exception.t_exception

| Res_attribute of Value.t_attribute

| Res_method of Value.t_method

| Res_section of string * text

| Res_recfield of Type.t_type * Type.record_field

| Res_const of Type.t_type * Type.variant_constructor

type search_result = result_element list

The type representing a research result.

val search_by_name : Module.t_module list -> Str.regexp -> search_result

Research of the elements whose name matches the given regular expression.

val values : Module.t_module list -> Value.t_value list

A function to search all the values in a list of modules.

val extensions : Module.t_module list -> Extension.t_extension_constructor list

A function to search all the extensions in a list of modules.

val exceptions : Module.t_module list -> Exception.t_exception list

A function to search all the exceptions in a list of modules.

val types : Module.t_module list -> Type.t_type list

A function to search all the types in a list of modules.

val attributes : Module.t_module list -> Value.t_attribute list

A function to search all the class attributes in a list of modules.

val methods : Module.t_module list -> Value.t_method list

A function to search all the class methods in a list of modules.

val classes : Module.t_module list -> Class.t_class list

A function to search all the classes in a list of modules.

val class_types : Module.t_module list -> Class.t_class_type list

A function to search all the class types in a list of modules.

val modules : Module.t_module list -> Module.t_module list

A function to search all the modules in a list of modules.

val module_types : Module.t_module list -> Module.t_module_type list

A function to search all the module types in a list of modules.