Module Heap_snapshot.Series

type t

val read : path:string -> t

At present, the Trace.t associated with a Series.t cannot be garbage collected or freed. This should not be a problem, since the intention is that a post-processor reads the trace and outputs another format.

val time_of_writer_close : t -> float

val num_threads : t -> int

type trace_kind =

| Normal

| Finaliser

val trace : t -> kind:trace_kind -> thread_index:int -> Trace.t option

val frame_table : t -> Frame_table.t

val shape_table : t -> Shape_table.t

val num_snapshots : t -> int

val snapshot : t -> index:int -> heap_snapshot

val events : t -> Event.t list

val has_call_counts : t -> bool

Returns true iff call count information was recorded in the series.