Module Indirect_call_point.Callee

type t

val callee : t -> Function_entry_point.t

The address of the first instruction of the callee.

val callee_node : t -> node

The node corresponding to the callee.

val call_count : t -> int option

The number of times the callee was called. This returns None in the same circumstances as Direct_call_point.call_count, above.

val next : t -> t option

Move to the next callee to which this call point has branched. None is returned when the end of the list is reached.