Module OCaml.Node

type t = ocaml_node

A node corresponding to an invocation of a function written in OCaml.

val compare : t -> t -> int

val function_identifier : t -> Function_identifier.t

A unique identifier for the function corresponding to this node.

val next_in_tail_call_chain : t -> t

This function traverses a circular list.

val fields : t -> shape_table:Shape_table.t -> Field.t option