Module Ephemeron.K2

Ephemerons with two keys.

type ('k1, 'k2, 'd) t

an ephemeron with two keys

val create : unit -> ('k1, 'k2, 'd) t

Same as Ephemeron.K1.create

alert old_ephemeron_api This function won't be available in 5.0

val get_key1 : ('k1, 'k2, 'd) t -> 'k1 option

Same as Ephemeron.K1.get_key

alert old_ephemeron_api This function won't be available in 5.0

val get_key1_copy : ('k1, 'k2, 'd) t -> 'k1 option

Same as Ephemeron.K1.get_key_copy

alert old_ephemeron_api This function won't be available in 5.0

val set_key1 : ('k1, 'k2, 'd) t -> 'k1 -> unit

Same as Ephemeron.K1.set_key

alert old_ephemeron_api This function won't be available in 5.0

val unset_key1 : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_key

alert old_ephemeron_api This function won't be available in 5.0

val check_key1 : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_key

alert old_ephemeron_api This function won't be available in 5.0

val get_key2 : ('k1, 'k2, 'd) t -> 'k2 option

Same as Ephemeron.K1.get_key

alert old_ephemeron_api This function won't be available in 5.0

val get_key2_copy : ('k1, 'k2, 'd) t -> 'k2 option

Same as Ephemeron.K1.get_key_copy

alert old_ephemeron_api This function won't be available in 5.0

val set_key2 : ('k1, 'k2, 'd) t -> 'k2 -> unit

Same as Ephemeron.K1.set_key

alert old_ephemeron_api This function won't be available in 5.0

val unset_key2 : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_key

alert old_ephemeron_api This function won't be available in 5.0

val check_key2 : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_key

alert old_ephemeron_api This function won't be available in 5.0

val blit_key1 : ('k1, _, _) t -> ('k1, _, _) t -> unit

Same as Ephemeron.K1.blit_key

alert old_ephemeron_api This function won't be available in 5.0

val blit_key2 : (_, 'k2, _) t -> (_, 'k2, _) t -> unit

Same as Ephemeron.K1.blit_key

alert old_ephemeron_api This function won't be available in 5.0

val blit_key12 : ('k1, 'k2, _) t -> ('k1, 'k2, _) t -> unit

Same as Ephemeron.K1.blit_key

alert old_ephemeron_api This function won't be available in 5.0

val get_data : ('k1, 'k2, 'd) t -> 'd option

Same as Ephemeron.K1.get_data

alert old_ephemeron_api This function won't be available in 5.0

val get_data_copy : ('k1, 'k2, 'd) t -> 'd option

Same as Ephemeron.K1.get_data_copy

alert old_ephemeron_api This function won't be available in 5.0

val set_data : ('k1, 'k2, 'd) t -> 'd -> unit

Same as Ephemeron.K1.set_data

alert old_ephemeron_api This function won't be available in 5.0

val unset_data : ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.unset_data

alert old_ephemeron_api This function won't be available in 5.0

val check_data : ('k1, 'k2, 'd) t -> bool

Same as Ephemeron.K1.check_data

alert old_ephemeron_api This function won't be available in 5.0

val blit_data : ('k1, 'k2, 'd) t -> ('k1, 'k2, 'd) t -> unit

Same as Ephemeron.K1.blit_data

alert old_ephemeron_api This function won't be available in 5.0

val make : 'k1 -> 'k2 -> 'd -> ('k1, 'k2, 'd) t

Same as Ephemeron.K1.make

val query : ('k1, 'k2, 'd) t -> 'k1 -> 'k2 -> 'd option

Same as Ephemeron.K1.query

module Make (H1 : Hashtbl.HashedType) (H2 : Hashtbl.HashedType) : S with type key = H1.t * H2.t

Functor building an implementation of a weak hash table

module MakeSeeded (H1 : Hashtbl.SeededHashedType) (H2 : Hashtbl.SeededHashedType) : SeededS with type key = H1.t * H2.t

Functor building an implementation of a weak hash table. The seed is similar to the one of Hashtbl.MakeSeeded.

module Bucket : sig ... end