Module Stdlib.StdLabels

Standard labeled libraries.

This meta-module provides versions of the Array, Bytes, List and String modules where function arguments are systematically labeled. It is intended to be opened at the top of source files, as shown below.

open StdLabels

let to_upper = String.map ~f:Char.uppercase_ascii
let seq len = List.init ~f:(function i -> i) ~len
let everything = Array.create_matrix ~dimx:42 ~dimy:42 42

module Array = ArrayLabels

module Bytes = BytesLabels

module List = ListLabels

module String = StringLabels