Module Topeval

This module provides two alternative implementations for internals of Toploop, for bytecode and native code.

You should not use it directly, refer to the functions in Toploop instead.